[libwin32darkmode] NppDarkMode_TabControl_drawItem(): Cast _tcslen() to int.
src\libwin32darkmode\NppDarkMode.cpp(145,37): warning C4267: 'argument':
conversion from 'size_t' to 'int', possible loss of data

+++ src/libwin32darkmode/NppDarkMode.cpp
@@ -142,7 +142,7 @@ static void NppDarkMode_TabControl_drawItem(HWND hWnd, const DRAWITEMSTRUCT *pDr
 
 	const COLORREF textColor = isSelected ? colorActiveText : colorInactiveText;
 	SetTextColor(hDC, textColor);
-	DrawText(hDC, decodedLabel, _tcslen(decodedLabel), &rect, Flags);
+	DrawText(hDC, decodedLabel, static_cast<int>(_tcslen(decodedLabel)), &rect, Flags);
 	RestoreDC(hDC, nSavedDC);
 }
